[BREAKING] Move remaining lazy mode constructs to @glimmer/opcode-compiler
