+++ packages/@glimmer/opcode-compiler/lib/partial-template.ts
