+++ packages/@glimmer/opcode-compiler/package.json
@@ -3,6 +3,8 @@
   "version": "0.30.2",
   "repository": "https://github.com/glimmerjs/glimmer-vm/tree/master/packages/@glimmer/opcode-compiler",
   "dependencies": {
+    "@glimmer/reference": "^0.30.2",
+    "@glimmer/runtime": "^0.30.2",
     "@glimmer/syntax": "^0.30.2",
     "@glimmer/util": "^0.30.2",
     "@glimmer/wire-format": "^0.30.2",
