+++ packages/@glimmer/runtime/lib/compiled/opcodes/component.ts
@@ -1,6 +1,5 @@
 import { Op, Register } from '@glimmer/vm';
 import { TemplateMeta } from '@glimmer/wire-format';
-import { ATTRS_BLOCK } from '@glimmer/opcode-compiler';
 import { Opaque, Option, Dict, dict, assert, unreachable, expect } from '@glimmer/util';
 import {
   VMHandle,
@@ -439,7 +438,7 @@ APPEND_OPCODES.add(Op.InvokeComponentLayout, (vm, { op1: _state }) => {
     };
 
     let blocks = args.blocks;
-    bindBlock(ATTRS_BLOCK, 'attrs');
+    bindBlock('&attrs', 'attrs');
     bindBlock('&inverse', 'else');
     bindBlock('&default', 'main');
