+++ packages/@glimmer/runtime/test/ember-component-test.ts
@@ -15,8 +15,9 @@ import {
   TestModifierManager
 } from "@glimmer/test-helpers";
 import { assign } from "@glimmer/util";
-import { RenderResult, Template, clientBuilder } from '@glimmer/runtime';
+import { RenderResult, clientBuilder } from '@glimmer/runtime';
 import { assert } from './support';
+import { Template } from '@glimmer/opcode-compiler';
 
 export class EmberishRootView extends EmberObject {
   public element: Element;
